+++ htbulma/services/mbox.py
@@ -18,7 +18,7 @@ class MBox(TagBulma):
         self["info"]="mbox"
         parent.add( self )
 
-    def show(self, *content, canClose=True):
+    def show(self, content, canClose=True):
         self.clear()
         o = Tag.div(_class = "modal is-active")
 
@@ -34,7 +34,7 @@ class MBox(TagBulma):
         else:
             o <= Tag.div(_class="modal-background")
 
-        o <= Tag.div( Box(*content), _class="modal-content")
+        o <= Tag.div( Box(content), _class="modal-content")
         self.add( o )
 
     def confirm(self, content, ok, ko=None,txtok="OK",txtko="Cancel"):
@@ -68,7 +68,7 @@ if __name__=="__main__":
         print("ko")
 
     obj=Tag( )
-    # MBox(obj).show( Tag.h1("titre"),Content("YO"), Button("nothing") )
+    # MBox(obj).show( Content("YO") )
     MBox(obj).confirm("hello ? sure ?????", ok=majok, ko=majko)
 
     from .. import _test